Move page reorder view to wagtail.admin.views.pages.ordering
