+++ wagtail/admin/views/pages/ordering.py
@@ -1,0 +1,44 @@
+from django.core.exceptions import PermissionDenied
+from django.http import HttpResponse
+from django.shortcuts import get_object_or_404
+
+from wagtail.core.models import Page
+
+
+def set_page_position(request, page_to_move_id):
+    page_to_move = get_object_or_404(Page, id=page_to_move_id)
+    parent_page = page_to_move.get_parent()
+
+    if not parent_page.permissions_for_user(request.user).can_reorder_children():
+        raise PermissionDenied
+
+    if request.method == 'POST':
+        # Get position parameter
+        position = request.GET.get('position', None)
+
+        # Find page thats already in this position
+        position_page = None
+        if position is not None:
+            try:
+                position_page = parent_page.get_children()[int(position)]
+            except IndexError:
+                pass  # No page in this position
+
+        # Move page
+
+        # any invalid moves *should* be caught by the permission check above,
+        # so don't bother to catch InvalidMoveToDescendant
+
+        if position_page:
+            # If the page has been moved to the right, insert it to the
+            # right. If left, then left.
+            old_position = list(parent_page.get_children()).index(page_to_move)
+            if int(position) < old_position:
+                page_to_move.move(position_page, pos='left')
+            elif int(position) > old_position:
+                page_to_move.move(position_page, pos='right')
+        else:
+            # Move page to end
+            page_to_move.move(parent_page, pos='last-child')
+
+    return HttpResponse('')
